+++ application/modules/question/views/admin/index.php
@@ -20,8 +20,10 @@
 		<td><a href="question/admin/question_detail/index/<?php echo $item['id']; ?>"><?php echo $item['name'] ?></a></td>
 		<td><?php echo $item['cnt'] ?></td>
 		<td><?php echo $item['userfirstname'] ?> <?php echo $item['usersurname'] ?></td>
-		<td><a href="question/admin/question/form/<?php echo $item['id'] ?>" class="btn" title="แก้ไข">แก้ไข</a>
-				 <a href="question/admin/question/delete/<?php echo $item['id'] ?>" class="btn" title="ลบ">ลบ</a></td>
+		<td><a href="question/admin/question/form/<?php echo $item['id'] ?>"   class="btn" title="แก้ไข">แก้ไข</a>
+			<a href="question/admin/question/delete/<?php echo $item['id'] ?>" class="btn" title="ลบ" onclick="return confirm('<?php echo NOTICE_CONFIRM_DELETE?>')">ลบ</a></td>
 	</tr>
 	<?php endforeach; ?>
 </table>
+
+
